Change key_len parameter to consistently be size_t
This avoids the risk of silent signed/unsigned type conversions.

(cherry picked from commit 926d54fafddcb0a513e034e2019e05507b9b7e13)

+++ src/handler/lib/get_merkleized_map_value.c
@@ -8,9 +8,9 @@
 int call_get_merkleized_map_value(dispatcher_context_t *dispatcher_context,
                                   const merkleized_map_commitment_t *map,
                                   const uint8_t *key,
-                                  int key_len,
+                                  size_t key_len,
                                   uint8_t *out,
-                                  int out_len) {
+                                  size_t out_len) {
     // LOG_PROCESSOR(__FILE__, __LINE__, __func__);
 
     uint8_t key_merkle_hash[32];
